+++ src/main/java/de/maxhenkel/camera/integration/waila/PluginCar.java
@@ -1,0 +1,19 @@
+package de.maxhenkel.camera.integration.waila;
+
+import de.maxhenkel.camera.entities.ImageEntity;
+import mcp.mobius.waila.api.IRegistrar;
+import mcp.mobius.waila.api.IWailaPlugin;
+import mcp.mobius.waila.api.TooltipPosition;
+import mcp.mobius.waila.api.WailaPlugin;
+
+@WailaPlugin
+public class PluginCar implements IWailaPlugin {
+
+    @Override
+    public void register(IRegistrar registrar) {
+        registrar.registerComponentProvider(HUDHandlerCars.INSTANCE, TooltipPosition.HEAD, ImageEntity.class);
+        registrar.registerComponentProvider(HUDHandlerCars.INSTANCE, TooltipPosition.BODY, ImageEntity.class);
+        registrar.registerComponentProvider(HUDHandlerCars.INSTANCE, TooltipPosition.TAIL, ImageEntity.class);
+    }
+
+}